Update README.md
I ran into issues when trying to run the application in Ubuntu 24.04 but I eventually got it running. I updated the readme to include the steps I took in case someone else could find it useful. I also updated the git command to reflect the repo's current URL and name.